+++ ChangeLog.md
@@ -1,5 +1,7 @@
 # v0.6.0 (?)
 * Added: The contrast sensitivity is computed separately for test and reference images. This much improves the predictions of display quality and for tone mapping. See Sec. 3.3 in [this paper](http://dx.doi.org/10.1145/3799902.3811107).
+* Added: `heatmap_file` passed to the constructor of the `cvvdp` class allows to save heatmaps to video files (of infinite length)
+* Added: new examples - `ex_luminance_masking.py`, `ex_contrast_masking.py`, `ex_custom_video_source.py`, `ex_noise_concealment.py`
 
 # v0.5.6 (4/06/2026)
 * Bug fix: incorrect assert in dm_preview_metric.py